% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_events.R
\name{get_events}
\alias{get_events}
\title{Fetch events for a specified season_id using StatsBombR}
\usage{
get_events(matches)
}
\arguments{
\item{matches}{A tibble. The output of `get_matches`.}
}
\value{
List. A list of metadata and tibbles. One tibble per match, containing all event 
  data for that match.
}
\description{
`get_events()` downloads events data using StatsBombr package.
}
