% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_shots.R
\name{get_shots}
\alias{get_shots}
\title{Extract and process shot events from StatsBomb data}
\usage{
get_shots(event)
}
\arguments{
\item{events}{A list of tibbles or a single tibble of Statsbomb events, the
output of `get_events()`.}
}
\value{
A list of tibbles. One tibble per match, containing shot data for 
that match.
}
\description{
`get_shots()` takes the output from `get_events()` and processes,
shot events for analysis or visualisation.
}
